requirements.txt
langchain-openai
langchain-text-splitters
pydantic-ai-slim[openai,logfire]
python-dotenv
neo4j
neomodel
simple-file-checksum
tree-sitter
tree-sitter-javascript
ollama
fastapi
uvicorn
serpapi
psycopg2-binary